requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.0
bs4==0.0.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
DateTime==4.3
dialogflow==0.6.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.17.1
futures==3.1.1
fuzzywuzzy==0.17.0
google-api-core==1.14.2
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.6.0
goslate==1.5.1
grpcio==1.24.0
gunicorn==19.9.0
httplib2==0.13.0
idna==2.8
inflect==2.1.0
isort==4.3.21
itsdangerous==1.1.0
jaraco.itertools==4.4.2
Jinja2==2.10.1
lazy-object-proxy==1.4.3
line-bot-sdk==1.12.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
newsapi-python==0.2.5
numpy==1.17.2
numpy-stl==2.10.1
oauth2client==4.1.3
Pillow==6.2.1
protobuf==3.9.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyDrive==1.3.1
pylint==2.4.3
python-resize-image==1.1.19
python-utils==2.3.0
pytz==2019.1
PyYAML==5.1.1
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.9.3
SQLAlchemy==1.3.5
stl==0.0.3
translation==1.0.5
uritemplate==3.0.0
urllib3==1.24.3
virtualenv==16.6.2
visitor==0.1.3
Werkzeug==0.16.0
wikipedia==1.4.0
wolframalpha==3.0.1
wrapt==1.11.2
WTForms==2.2.1
xmltodict==0.12.0
yandex-translater==6.0
zope.interface==4.6.0
